IconText now exists. Image changes, and image style changes.
IconText shows an icons title (used only for icons without a link property)

Images for Repopit
and Operation portal have been changed. Image styles are no longer mandatory for 100% width and will adjust width to be more suitable for the image.
