+++ client/src/components/main-content/Projects.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Grid } from '@material-ui/core';
-import CoverImg from '../../images/website-wireframe.png';
-import OperationPortal from '../../images/operation-portal.png';
+import Repopit from '../../images/repopit-dev-cover.png';
+import OperationPortal from '../../images/operation-portal-cover.png';
 import FocusPocus from '../../images/focus-pocus-cover.png';
 import AntiBody from '../../images/anti-body.png';
 import { SectionContainer, SectionHeader, SectionItem, SectionDescription, LinkWrapper as LW} from '../sub-components/index.js';
-import { HtmlIcon, JsIcon, CssIcon, ReactIcon, ReduxIcon, PostgreSqlIcon, DotNetIcon, FlutterIcon, MaterialUiIcon, AwsIcon, GithubIcon, WebIcon, AndroidIcon, IosIcon, MongoDbIcon, ExpressIcon, NodeJsIcon, CSharpIcon, UnityIcon } from '../../SvgIcons.js';
+import { HtmlIcon, JsIcon, CssIcon, ReactIcon, ReduxIcon, PostgreSqlIcon, DotNetIcon, FlutterIcon, MaterialUiIcon, AwsIcon, GithubIcon, WebIcon, AndroidIcon, IosIcon, MongoDbIcon, ExpressIcon, NodeJsIcon, CSharpIcon, UnityIcon } from '../SvgIcons.js';
 
 export default function Projects() {
     return (
@@ -13,7 +13,7 @@ export default function Projects() {
             <Grid container item direction="column" justify="flex-start" alignItems="center">
                 <SectionContainer position={1}>
                     <h2><SectionHeader headerTitle={"Latest Development"} linkDestination={"/fullstack"}/></h2>
-                    <SectionItem img={CoverImg} title={"Repopit.dev"} internalLink={"/projects/repopit"}>
+                    <SectionItem img={Repopit} title={"Repopit.dev"} internalLink={"/projects/repopit"}>
                         <SectionDescription title={"Repopit.dev"} contributors={["Trevor Bland"]} 
                         technologies={<span><HtmlIcon /><JsIcon /><CssIcon /><ReactIcon /><ReduxIcon /><MaterialUiIcon /><AwsIcon /></span>} 
                         liveLinks={<LW link={"https://repopit.dev"}><WebIcon /></LW>} sourceLink={<LW link={"portfolio-web"}><GithubIcon /></LW>}
